requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
google-genai>=1.30.0
gunicorn>=23.0.0
